requirements.txt
amqp==2.1.4
asgi-redis==1.4.0
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
billiard==3.5.0.2
bleach==2.0.0
celery==4.0.2
certifi==2017.4.17
channels==1.1.3
chardet==3.0.4
constantly==15.1.0
daphne==1.2.0
Django==1.11.2
django-celery-beat==1.0.1
django-channels-presence==0.0.7
django-widget-tweaks==1.4.1
html5lib==0.999999999
hyperlink==17.1.1
idna==2.5
incremental==17.5.0
kombu==4.0.2
Markdown==2.6.8
msgpack-python==0.4.8
pkg-resources==0.0.0
postmarker==0.11.2
psycopg2==2.7.1
pytz==2017.2
redis==2.10.5
requests==2.17.3
six==1.10.0
Twisted==17.5.0
txaio==2.8.0
urllib3==1.21.1
vine==1.1.3
webencodings==0.5.1
zope.interface==4.4.1
